# backend/data/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/admin/user/list":{"get":{"operationId":"UserAdminController_list_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["modules.admin.user"]}},"/api/v1/admin/user/create":{"post":{"operationId":"UserAdminController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["modules.admin.user"]}},"/api/v1/admin/user/update/unblocked":{"patch":{"operationId":"UserAdminController_unblockedBulk_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBulkRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["modules.admin.user"]}},"/api/v1/admin/user/update/blocked":{"patch":{"operationId":"UserAdminController_blockedBulk_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBulkRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["modules.admin.user"]}},"/api/v1/admin/user/delete":{"delete":{"operationId":"UserAdminController_deleteBulk_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBulkRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["modules.admin.user"]}},"/api/v1/public/user/login":{"post":{"operationId":"UserPublicController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["modules.public.user"]}},"/api/v1/public/user/sign-up":{"post":{"operationId":"UserPublicController_signUp_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["modules.public.user"]}},"/api/v1/auth/user/refresh":{"post":{"operationId":"UserAuthController_refresh_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["modules.auth.user"]}}},"info":{"title":"userControlDashboard APIs Specification","description":"Section for describe whole APIs","version":"1.0","contact":{}},"tags":[{"name":"API's","description":""}],"servers":[{"url":"/"},{"url":"/staging"},{"url":"/prod"}],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"refreshToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserCreateDto":{"type":"object","properties":{"email":{"type":"string","example":"Lacey_Donnelly63@hotmail.com"},"firstName":{"type":"string","example":"Gerardo"},"lastName":{"type":"string","example":"Gusikowski"},"password":{"type":"string","description":"string password","example":"o}@@!123"}},"required":["email","firstName","lastName","password"]},"UserBulkRequestDto":{"type":"object","properties":{}},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"Lacey_Donnelly63@hotmail.com"},"password":{"type":"string","description":"string password","example":"o}@@!123"}},"required":["email","password"]},"UserSignUpDto":{"type":"object","properties":{"email":{"type":"string","example":"Lacey_Donnelly63@hotmail.com"},"firstName":{"type":"string","example":"Gerardo"},"lastName":{"type":"string","example":"Gusikowski"},"password":{"type":"string","description":"string password","example":"o}@@!123"}},"required":["email","firstName","lastName","password"]}}}}